Refactor the Aes encryption implementation to reuse the same design as the TripleDes implementation.
This cuts down on a lot of duplicate code, since we can reuse the UniversalCryptoTransform classes.
