+++ src/System.Security.Cryptography.Algorithms/src/Internal/Cryptography/AesImplementation.Windows.cs
@@ -1,34 +1,56 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
+using System;
 using System.Security.Cryptography;
+using Internal.NativeCrypto;
 
 namespace Internal.Cryptography
 {
     internal partial class AesImplementation
     {
-        private static ICryptoTransform CreateEncryptor(
+        private static ICryptoTransform CreateTransformCore(
             CipherMode cipherMode,
             PaddingMode paddingMode,
             byte[] key,
             byte[] iv,
-            int blockSize)
+            int blockSize,
+            bool encrypting)
         {
-            return new AesCngCryptoEncryptor(cipherMode, paddingMode, key, iv, blockSize);
-        }
+            SafeAlgorithmHandle algorithm = GetCipherAlgorithm(cipherMode);
 
-        private static ICryptoTransform CreateDecryptor(
-            CipherMode cipherMode,
-            PaddingMode paddingMode,
-            byte[] key,
-            byte[] iv,
-            int blockSize)
-        {
-            return new AesCngCryptoDecryptor(cipherMode, paddingMode, key, iv, blockSize);
+            BasicSymmetricCipher cipher = new CngCipher(algorithm, cipherMode, blockSize, key, iv, encrypting);
+            return UniversalCryptoTransform.Create(paddingMode, cipher, encrypting);
         }
 
         // -----------------------------
         // ---- PAL layer ends here ----
         // -----------------------------
+
+        private static readonly SafeAlgorithmHandle s_hAlgCbc = OpenAesAlgorithm(Cng.BCRYPT_CHAIN_MODE_CBC);
+        private static readonly SafeAlgorithmHandle s_hAlgEcb = OpenAesAlgorithm(Cng.BCRYPT_CHAIN_MODE_ECB);
+
+        private static SafeAlgorithmHandle GetCipherAlgorithm(CipherMode cipherMode)
+        {
+            // Windows 8 added support to set the CipherMode value on a key,
+            // but Windows 7 requires that it be set on the algorithm before key creation.
+            switch (cipherMode)
+            {
+                case CipherMode.CBC:
+                    return s_hAlgCbc;
+                case CipherMode.ECB:
+                    return s_hAlgEcb;
+                default:
+                    throw new NotSupportedException();
+            }
+        }
+
+        private static SafeAlgorithmHandle OpenAesAlgorithm(string cipherMode)
+        {
+            SafeAlgorithmHandle hAlg = Cng.BCryptOpenAlgorithmProvider(Cng.BCRYPT_AES_ALGORITHM, null, Cng.OpenAlgorithmProviderFlags.NONE);
+            hAlg.SetCipherMode(cipherMode);
+
+            return hAlg;
+        }
     }
 }
