+++ src/System.Security.Cryptography.Algorithms/src/Internal/Cryptography/CngCipher.cs
@@ -8,19 +8,18 @@ using Internal.NativeCrypto;
 
 namespace Internal.Cryptography
 {
-    internal sealed class TripleDesCngCipher : BasicSymmetricCipher
+    internal sealed class CngCipher : BasicSymmetricCipher
     {
-        private static readonly SafeAlgorithmHandle s_hAlgCbc = Open3DesAlgorithm(Cng.BCRYPT_CHAIN_MODE_CBC);
-        private static readonly SafeAlgorithmHandle s_hAlgEcb = Open3DesAlgorithm(Cng.BCRYPT_CHAIN_MODE_ECB);
-
         private readonly bool _encrypting;
         private SafeKeyHandle _hKey;
         private byte[] _currentIv;  // CNG mutates this with the updated IV for the next stage on each Encrypt/Decrypt call.
                                     // The base IV holds a copy of the original IV for Reset(), until it is cleared by Dispose().
 
-        public TripleDesCngCipher(CipherMode cipherMode, int blockSizeInBytes, byte[] key, byte[] iv, bool encrypting)
+        public CngCipher(SafeAlgorithmHandle algorithm, CipherMode cipherMode, int blockSizeInBytes, byte[] key, byte[] iv, bool encrypting)
             : base(cipherMode.GetCipherIv(iv), blockSizeInBytes)
         {
+            Debug.Assert(algorithm != null);
+
             _encrypting = encrypting;
 
             if (IV != null)
@@ -28,8 +27,7 @@ namespace Internal.Cryptography
                 _currentIv = new byte[IV.Length];
             }
 
-            SafeAlgorithmHandle hAlg = GetCipherAlgorithm(cipherMode);
-            _hKey = hAlg.BCryptImportKey(key);
+            _hKey = algorithm.BCryptImportKey(key);
             Reset();
         }
 
@@ -112,30 +110,6 @@ namespace Internal.Cryptography
             {
                 Buffer.BlockCopy(IV, 0, _currentIv, 0, IV.Length);
             }
-        }
-
-        private static SafeAlgorithmHandle GetCipherAlgorithm(CipherMode cipherMode)
-        {
-            // Windows 8 added support to set the CipherMode value on a key,
-            // but Windows 7 requires that it be set on the algorithm before key creation.
-            switch (cipherMode)
-            {
-                case CipherMode.CBC:
-                    return s_hAlgCbc;
-                case CipherMode.ECB:
-                    return s_hAlgEcb;
-                default:
-                    // This is what AesCngCryptoTransform::GetCipherAlgorithm throws when it doesn't understand the value.
-                    throw new NotSupportedException();
-            }
-        }
-
-        private static SafeAlgorithmHandle Open3DesAlgorithm(string cipherMode)
-        {
-            SafeAlgorithmHandle hAlg = Cng.BCryptOpenAlgorithmProvider(Cng.BCRYPT_3DES_ALGORITHM, null, Cng.OpenAlgorithmProviderFlags.NONE);
-            hAlg.SetCipherMode(cipherMode);
-
-            return hAlg;
         }
     }
 }
