+++ src/System.Security.Cryptography.Algorithms/src/Internal/Cryptography/TripleDesImplementation.Unix.cs
@@ -1,13 +1,14 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
+using System;
 using System.Security.Cryptography;
 
 namespace Internal.Cryptography
 {
     partial class TripleDesImplementation
     {
-        private ICryptoTransform CreateTransformCore(
+        private static ICryptoTransform CreateTransformCore(
             CipherMode cipherMode,
             PaddingMode paddingMode,
             byte[] key,
@@ -15,7 +16,21 @@ namespace Internal.Cryptography
             int blockSize,
             bool encrypting)
         {
-            BasicSymmetricCipher cipher = new TripleDesOpenSslCipher(cipherMode, blockSize, key, iv, encrypting);
+            // The algorithm pointer is a static pointer, so not having any cleanup code is correct.
+            IntPtr algorithm;
+            switch (cipherMode)
+            {
+                case CipherMode.CBC:
+                    algorithm = Interop.Crypto.EvpDes3Cbc();
+                    break;
+                case CipherMode.ECB:
+                    algorithm = Interop.Crypto.EvpDes3Ecb();
+                    break;
+                default:
+                    throw new NotSupportedException();
+            }
+
+            BasicSymmetricCipher cipher = new OpenSslCipher(algorithm, cipherMode, blockSize, key, iv, encrypting);
             return UniversalCryptoTransform.Create(paddingMode, cipher, encrypting);
         }
